requirements.txt
torch==2.3.1
torchvision==0.18.1
numpy==1.26.4
PyYAML==6.0.2